% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{accuracy}
\alias{accuracy}
\title{Miscelaneous function used in main routines.}
\usage{
accuracy(testdata, fitteddata)
}
\arguments{
\item{testdata}{a vector of occurrences.}

\item{fitteddata}{a vector of occurrences.}
}
\value{
A numeric indicating the degree of correspondence.
}
\description{
Miscelaneous function used in main routines.
}
\examples{
observed <- c(1,1,1,0,0,0,0,0,0,0)
predicted <-c(1,0,1,0,0,0,0,1,1,0)

print(accuracy(observed, predicted))
}

