% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{autoGLM}
\alias{autoGLM}
\title{A main function of the package for automated generalized linear model fitting. Includes additional options w.r.t. the \code{\link{generalizeToSpecific}} command.}
\usage{
autoGLM(data, reclasstable = "default", class = 1,
  outputpath = paste(getwd(), "//", sep = ""), modelname = "autoGLM",
  tracelevel = 1, actions = c("print", "return"), NAval = -9999,
  model = "logit", preselect = "lm", method = "opt.ic", crit.t = 1.64,
  crit.p = 0.05, test = "LR", KLIC = "AICc", accuracytolerance = 0.01,
  confidence.alternative = 0.9, use.share = 0.25, maxsampleruns = 50,
  memorymanagement = TRUE, returnall = FALSE, compress = FALSE,
  JIT = TRUE)
}
\arguments{
\item{data}{A dataframe with a categorical response variable in the first column, and covariates in subsequent columns. Typically the product of cbind(Y,X).}

\item{reclasstable}{A table that maps the first column of data into a binary response variable. By default it will be ommitted (the binary response variable will be identical to data[,1]). See also \code{\link{corinetable}}, See also \code{\link{reclassify}}.}

\item{class}{The class that should be 1 in the binary response variable, all other classes in the categorical variable will be set to 0. Defaults to 1. See also \code{\link{reclassify}}.}

\item{outputpath}{The location on the hard drive where output wwill be written to. Defaults to getwd().}

\item{modelname}{The name of the model, will be used when writing a weightsfile. Defaults to "autoGLM". See also \code{\link{exportWeightsfile}}.}

\item{tracelevel}{The amount of information to be printed. Passed on to underlying routines. Defaults to 1 for printing, set to 0 for no printing.}

\item{actions}{Actions to be taken by autoGLM, by default c("print", "return"), may include any combination of c("write", "print", "log", "return"), for writing a geoDMS weightsfile, See also \code{\link{exportWeightsfile}}, printing results, writing a log file, and returning results as a list object.}

\item{NAval}{Optional categorical variable that should be dropped by the reclassification scheme. See also \code{\link{reclassify}}.}

\item{model}{Main model type that should be calibrated, either "lm", "probit", or "logit". See also \code{\link{generalizeToSpecific}}.}

\item{preselect}{Optional variable preselection using a first order approximation (linear model) of the logit or probit model, by specifying "lm" (default setting). See also \code{\link{selectX}}.}

\item{method}{The optimization strategy. Either "opt.ic" to optimize using information criteria, "opt.t" for step-wise elimination of insignificant values
(statistically speaking not a sound procedure, but it will provide a parsimonious model that can be usefull as a benchmark), or "opt.h" to optimize by classical hypothesis tests.
defaults to "opt.ic". See also See also \code{\link{opt.ic}}, \code{\link{opt.t}}, See also \code{\link{opt.h}}.}

\item{crit.t}{The t-value indicating significance when using method "opt.t", defaults to 1.64. \code{\link{opt.t}}.}

\item{crit.p}{the p-value used by method "opt.h" in the hypothesis tests. Defaults to 0.05. \code{\link{opt.h}}.}

\item{test}{The hypothesis test used by "opt.h". Defaults to "LR" for the Likelihood Ratio test. Other options are "F", for an F test for joint significance of insignificant parameters, or "Chisq" for a wald test against the Chi squared distribution. \code{\link{opt.h}}.}

\item{KLIC}{The information criterion used by "opt.ic", either "AIC" or "AICc", defaults to the latter. \code{\link{opt.ic}}.}

\item{accuracytolerance}{When aut of sample and within sample accuracy differ more than accuracytolerance, a warning will be issued, which is also logged when specifying "log" in actions. Defaults to 0.01. \code{\link{accuracy}}.}

\item{confidence.alternative}{See also \code{\link{getSamples}}, confidence level used for the alternative of dissimilar samples in the sampling routine. Defaults to .85.}

\item{use.share}{Share of the data used, See also \code{\link{getSamples}}. Defaults to .25.}

\item{maxsampleruns}{See also \code{\link{getSamples}}, defaults to 50.}

\item{memorymanagement}{TRUE/FALSE indicating whether garbage collection should be forced regularly when memory usage is high. Defaults to TRUE, recommended setting for large datasets. See also \code{\link{tgc}}.}

\item{returnall}{TRUE, FALSE, or "writedisk" indicating whether all the outputted objects for each class should be returned in an array as produced by lapply, or whether only the final output should be returnd as an object.
Specifying "writedisk" will write the objects containing results of each class as seperate .RDS files, which you can use to restore the output using readRDS(). \code{\link{iapply}}.
Returning an array of all results can consume large amounts of memory as each object contains copies of the used datasets. When working with countrysize datasets, these array objects can easily require over 64gb of RAM.
Specifying returnall = FALSE (default setting), is much more more RAM friendly as it stores results for each class in the same memory adress, overwriting previous results. Seting returnall = FALSE, will still write log files and print diagnostics to screen if specified in actions.
returnall="writedisk" is the recommended setting, but it is not default. \code{\link{iapply}}.}

\item{compress, }{passed on to iapply. Defaults to no compression of RDS output, which is the recommended setting if computation time is valued of disk space. Keep in mind that when using large datasets, autoGLM objects can be several gigabytes in size. \code{\link{iapply}}.}

\item{JIT, }{logical indicating whether just-in-time compilation of internal functions should be used. Mainly for historical reasons.}
}
\description{
This function is a wrapper around the optimization and selection routines in the package and can be used for automated calibration of GLM's on semi large datasets.
\code{\link{generalizeToSpecific}} is more appropiate for manual R sessions, \code{\link{autoGLM}} is more appropiate for situations when calibration takes a long time.
E.g., it allows to run \code{\link{generalizeToSpecific}} over a vector of dependent variable classes, and to log and write outputs to disk.
}
\examples{

data(ITdata)
datacorinetable)

results <- autoGLM(data=randomlogit, reclasstable=corinetable, class=0, method ="opt.ic")


# All options:
autoGLM <- function (data, reclasstable = "default", class=1, outputpath=paste(getwd(),"//", sep=""),
 modelname="autoGLM", tracelevel=1,
 actions = c("print", "return"), NAval = -9999,
 model="logit", preselect = "lm", method = "opt.ic", crit.t = 1.64, crit.p =.05,
 test = "LR", KLIC = "AICc", accuracytolerance =0.01, confidence.alternative =0.90,
 use.share = 0.25, maxsampleruns=50, memorymanagement = TRUE, returnall = FALSE,
 compress = FALSE, JIT = TRUE)

}

