% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\docType{data}
\name{corinetable}
\alias{corinetable}
\title{A reclassification table for Corine land cover to a simplified aggregation used in the LUISA platform of the JRC.}
\usage{
data(corinetable)
}
\description{
A table that maps 51 Corine Land-cover classes into 8 classes and a NoData class.
\itemize{
	\item 0. Urban.
	\item 1. Industry.
	\item 2. Arable.
	\item 3. Permanent crops.
	\item 4. Pastures.
	\item 5. Forest.
	\item 6. Herbaceous cover.
	\item 16. Transitional.
	\item -9999. NoData.
}
}
\keyword{datasets}

