% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{describe}
\alias{describe}
\title{A simple function to describe a dataset.}
\usage{
describe(X)
}
\arguments{
\item{X}{Matrix or dataframe to be summarized.}
}
\value{
A dataframe with names and statistics.
}
\description{
This function returns a dataframe with variable names, their minimum and maximum values, means and standard deviations.
}
\examples{
someVector = 1:10
print(describe(someVector))

df <- data(ITdata)
description <- describe(df)
print(description)
}

