% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{drop.na.model}
\alias{drop.na.model}
\title{Miscelaneous function used in main routines.}
\usage{
drop.na.model(namodel)
}
\arguments{
\item{namodel}{modelobject.}
}
\value{
a model object of the same type as the supplied object.
}
\description{
If a model object produced by lm() or glm() has NA standard errors, drop.na.model() drops the problematic variables and refits the model using a warmstart, See also \code{\link{warmstart}}.
}

