% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{easygmmlogit}
\alias{easygmmlogit}
\title{A function for user-friendly estimation of logit models using GMM. Advanced users may use the more elaborate gmm package instead.}
\usage{
easygmmlogit(Y, X, start = "default", maximizer = "nlminb")
}
\arguments{
\item{Y}{A binary response variable.}

\item{X}{A dataframe of multiple exogenous regressors.}

\item{start}{A vector of starting values to be used by the numerical optimizer. By default, the function calls guessStartVal get a fast initial estimate using an IWLS estimator. If that fails, for example due to bad data input, starting values will be set to zero.}

\item{maximizer}{The numerical procedure to be used. Either "BFGS" for fast optimization, or "nlminb" for reliable PORT routines. Defaults to "nlminb".}
}
\value{
A model generated by gmm()
}
\description{
This function automizes GMM estimation of the logit model. GMM estimation runs very heavy compared to the IWLS estimator used by the standard routines used in this package, but GMM is a more general approach and works under milder conditions regarding the uniqueness of a solution.
}
\examples{
# Load data
data(ITdata)
data(corinetable)
#Grab a sample, gmm runs heavy.
sample <- ITdata[getSamples(ITdata, share =.05),]
# Reclassify 
catITdata <- reclassify(sample, reclasstable = corinetable)
# create a binary response dataset.
Y <- MLtoBinomData(catITdata[,1], class =0)
X <- catITdata[,c(2:4)]
GMM <- easygmmlogit (Y, X) 
}

