% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{exportWeightsfile}
\alias{exportWeightsfile}
\title{A function to export a csv file containing the weights file in a format that can be loaded into geoDMS.}
\usage{
exportWeightsfile(model, originaldata, modeldata, coefnamelist, outdir,
  modelname, filename)
}
\arguments{
\item{model}{The model object from which the estimated coefficients should be extracted.}

\item{originaldata}{The complete dataset from which the model data has been extracted. Used by the function to determine which variables have a zero weight.}

\item{modeldata}{The dataset on which the model is estimated. Used by the function to determine which variables have a zero weight.}

\item{coefnamelist}{A character vector containing the weight names that should be printed in the weights file.}

\item{outdir}{The directory to which the weightsfile should be exported.}

\item{modelname}{A string containing the name of the country that should be printed in the weightsfile.}

\item{filename}{The name of the exported weightsfile.}
}
\value{
Any messages that may be printed by file(), writeLines() or close().
}
\description{
This function exports a csv file containing the names of the weights, coefficients and country name in a standardized format supported by geoDMS.
}
\examples{
logitpars <- c(-0.5, 0.4, -0.3, 0.2, -0.1, 0, 0, 0, 0, 0, 0, 0)
someData <- simulateLogit(nobs=2500, nxregs=12, pars=logitpars) 
useData <- someData[,c(1:6)]
someModel <- glm(formula(useData), family=binomial(link='logit'), data=useData)

exportWeightsfile(model = someModel, originaldata = someData, modeldata = useData, 
				coefnamelist = colnames(someData[-1,]), outdir = "C:\\\\Users\\\\", 
					modelname = "myfirstglm", filename = "important_result.csv") 

ITdata <- data(ITdata)
ITsample <- getSamples(data = ITdata, share = 0.05, confidence.alternative=0.90, max.iter =100)
sampledIT = ITdata[ITsample,]
reclIT <- reclassify(sampledIT)
urbanIT <- MLtoBinomData(reclIT, class=0)
trainY = urbanIT[,1]
trainX = urbanIT[,-1]
bestX <- selectX(trainY, trainX)
bestlogit <- glm(formula(cbind(trainY,trainX)), family=binomial(link='logit'), data=cbind(trainY,trainX))

exportWeightsfile(model = bestlogit, originaldata = ITdata, modeldata = cbind(trainY,trainX), 
				coefnamelist = colnames(trainX), outdir = "C:\\\\Users\\\\", 
					modelname = "IT", filename = "Urban_weights.csv") 
}

