% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{fitnew}
\alias{fitnew}
\title{Miscelaneous function used in main routines.}
\usage{
fitnew(newspec, model, newZ, newstart)
}
\arguments{
\item{newspec}{formula object}

\item{model}{Either "lm", "logit", or "probit".}

\item{newZ}{dataframe.}

\item{newstart}{vector of starting values.}
}
\value{
model object of requested type
}
\description{
Miscelaneous function used in main routines.
}

