% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{getCall}
\alias{getCall}
\title{Returns the name of a call as string.}
\usage{
getCall(level = -1)
}
\arguments{
\item{level}{0 returns "getCall", -1 returns the name of the function in which getCall(-1) is called etc.}
}
\value{
string
}
\description{
When called inside a function, returns the parent function's name.
}

