% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{getCall2}
\alias{getCall2}
\title{Returns the name of a call as string.}
\usage{
getCall2(level = -1)
}
\arguments{
\item{level}{0 returns "getCall2", -1 returns the name of the function in which getCall2(-1) is called etc.}
}
\value{
string
}
\description{
When called inside a function, returns the parent function's name.
}
\examples{
getCall2(0)

foo <- function(){
return(getCall2())
}
foo()

bar <- function (somevar){
foo <- function(){
	return(getCall2(-2))
}
return(foo())
}
}

