% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{getSamples}
\alias{getSamples}
\title{A function to extract small samples that maintain important characteristics of the population sample.}
\usage{
getSamples(data, share = 0.25, confidence.alternative = 0.9,
  max.iter = 50, tracelevel = 1, memorymanagement = TRUE)
}
\arguments{
\item{data}{The population data from which a sample needs to be taken.}

\item{share}{The size of the sample in terms of the share of the population data. Defaults to .25.}

\item{confidence.alternative}{The confidence level used in the F and t-tests defined as the probability level at which the alternative is accepted. For confidence.alternative = .9, we need less evidence to accept the alternative hypothesis that the samples are unequal than at confidence.alternative = .95, hence .90 is stricter than .95.}

\item{max.iter}{The maximum number of draws to be taken. The programm breaks either when a suitable sample is found or when max.iter is reached.}

\item{tracelevel}{Similar to a verbose statement. Should information be printed during execution? defaults to 1 for printing. set to 0 for no printing.}

\item{memorymanagement}{TRUE/FALSE indicating whether garbage collection should be forced using tgc(). Defaults to TRUE. Recommended setting for large datasets.}
}
\value{
A sample of the population dataset that has significantly similar means and variances, or a message indicating that no suitable dat
}
\description{
This function returns a sample extracted from the supplied population data, that has a similar distribution to the supplied population dataset. The function is called by guessStartval() to estimate inital values for numerical optimization procedures, but can also be used directly to reduce the sample size such that computationally intensive models can be estimated on a representative sample of an entire dataset. The function makes use of var.test() to compute an F test for the ratio of sample/population variance, and t.test() to compare their means.
}
\examples{
getSamples (data = ITdata, share = 0.025, confidence.alternative=0.90, max.iter =100)
}

