% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{guessStartVal}
\alias{guessStartVal}
\title{A function to efficiently obtain starting values for numerical optimization procedures. Used to initialize the "warm start" optimization routines in generalizeTospecific. The funciton itself uses a "warm start" algorithm over a growing dataset similar to a sieves estimator for an unbounded parameter space.}
\usage{
guessStartVal(Y, X, model, tracelevel = 1, memorymanagement = TRUE)
}
\arguments{
\item{Y}{A binary response variable.}

\item{X}{A dataset containing multiple exogenous regressors.}

\item{model}{The model for which starting values should be estimated. Either "logit" or "probit" for the logit or probit model, or "gmm_nlminb" for a logit model estimated with gmm using PORT routines (reliable) or "gmm_bfgs" using the BFGS algorithm (fast, but still very slow compared to option "logit").}

\item{tracelevel}{Whether information should be printed during execution. Defaults to 1 for printing, set to 0 for no printing.}

\item{memorymanagement}{TRUE/FALSE indicating whether garbage collection should be forec regularly when memory usage is high. Defaults to TRUE, recommended setting for large datasets.}
}
\value{
A vector of coefficients that can be passed on to numerical optimization algorithms.
}
\description{
This function is called by generalizeToSpecific(), but may also be called by users directly to obtain an initial gues of starting values to be passed on to easygmmlogit().
}
\examples{
data(ITdata)
data(corinetable)
#Grab a sample (optional).
sample <- ITdata[getSamples(ITdata, share =.05),]
# Reclassify 
catITdata <- reclassify(sample, reclasstable = corinetable)
# create a binary response dataset.
Y <- MLtoBinomData(catITdata[,1], class =0)
X <- catITdata[,c(2:4)]
initial <- guessStartVal(Y, X, model="logit")
}

