% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{IC}
\alias{IC}
\title{A simple function to extract information criteria from a model object.}
\usage{
IC(model, KLIC = "AICc", sigma.is.estimated = FALSE)
}
\arguments{
\item{model}{A model object that is supported by AIC(). For example produced by lm(), glm(), or speedglm().}

\item{KLIC}{either "AIC" for the AIC, or "AICc" for the corrected criterion. defaults to "AICc".}

\item{sigma.is.estimated}{boolean indicating whether variance has been estimated as part of the parameters (Likelihood), or not (LS), If FALSE, number of parameters is increased by 1 for comparison with MLE objects. Defaults to FALSE.
Should not impact the results when models of the same type are compared, but the option has been added to allow users to change the number of counted parameters when comparing MLE results with LS results.}
}
\value{
numeric value.
}
\description{
This function returns a Kullback-Leibler Information Criterion.
}
\examples{
somepars <- c(-0.5, 0.4, -0.3, 0.2, -0.1, 0, 0, 0)
someData <- simulateLogit(nobs=2500, pars=somepars)
someModel <- logit(someData)
IC(someModel, KLIC="AIC")
AIC(someModel) # compare
IC(someModel) # AICc
}

