% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{is.speed}
\alias{is.speed}
\title{Miscelaneous function used in main routines.}
\usage{
is.speed(model)
}
\arguments{
\item{model}{a modelobject.}
}
\value{
Boolean indicating whether supplied object is of class speedglm from package speedglm.
}
\description{
Miscelaneous function used in main routines.
}
\examples{
y=c(1,0,1,0,1,0)
x=c(0,0,1,1,1,0)
is.speed(lm(y~x))
}

