% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\docType{data}
\name{ITdata}
\alias{ITdata}
\title{A small sample dataset produced with \code{\link{getSamples}} with 157550 obersvations of 100x100 meter resolution Corine land cover data for Italy with various exogenous variables.}
\format{a data frame with 157550 rows and 27 columns}
\usage{
data(ITdata)
}
\description{
To inspect the data, describe(data(ITdata))
}
\keyword{datasets}

