% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{logistic}
\alias{logistic}
\title{The logistic function.}
\usage{
logistic(theta, data)
}
\arguments{
\item{theta}{A vector of coefficients.}

\item{data}{A dataframe of multiple exogenous regressors.}
}
\value{
A vector of values produced by a logistic formula under specified parameters and data.
}
\description{
Called by various routines in the package. May also be used to predict fitted conditional probabilities by supplying a set of variables and corresponding coefficients estimated from a logit model.
}
\examples{
predicted <- logistic(coef(logitmodel), logitmodel$model[,-1])
}

