% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{logit}
\alias{logit}
\title{Simple wrapper to estimate a binomial logit model by IWLS or GMM.}
\usage{
logit(Z, method = "IWLS", start = "default", maximizer = "nlminb",
  wmatrix = "ident", gmmtype = "iterative")
}
\arguments{
\item{Z}{dataset with dependent variable in the first column and explanatory variables in other columns.}

\item{method}{either IWLS to estimate a glm by Iterated Weighted Least Squares, or GMM to estimate with method of moments.}

\item{start}{starting values, defaults to vector of zeros for IWLS and \code{\link{guessStartVal}} for gmm.}

\item{maximizer, }{maximizer used for GMM, either "nlminb" for port routines or "BFGS".}

\item{wmatrix}{Defaults to "ident" for identity matrix weighting of the gmm estimator. Other option is "optimal" for inverse covariance weighting. See package "gmm" for details.}

\item{gmmtype, }{Either "twostep" for the two step GMM proposed by Hansen(1982),
and the "cue" and "iterative" for respectively the continuously and the
iteratively updated methods proposed by Hansen, Eaton et Yaron (1996). See package "gmm" for details. Defaults to "iterative".}
}
\value{
A logit model fitted by glm()
}
\description{
This function returns an estimated logit model.
}
\examples{
set.seed(10101)
Z <- simulateLogit(nobs=1000, pars =c(0.25,-0.5, 0.1, -0.1, 0, 0))

# Iterated Weighted Least Squares
test<- logit(Z)
# GMM using port routines and the results of the IWLS as starting values
test2<- logit(Z, method="gmm", start=coef(test))
# GMM using BFGS and starting at correct parameters
test3<- logit(Z, method="gmm", maximizer="BFGS", start =c(0, 0.25,-0.5, 0.1, -0.1, 0, 0))
test4<- logit(Z, method="gmm")
}
\keyword{biomial}
\keyword{estimation}
\keyword{logit}

