% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{MLtoBinomData}
\alias{MLtoBinomData}
\title{Miscelaneous function used in main routines.}
\usage{
MLtoBinomData(data, class)
}
\arguments{
\item{data}{factorial or numeric data of a categorial nature.}

\item{class}{the category that should be reclassified into 1, all other values return as 0.}
}
\value{
a binary dataset.
}
\description{
Miscelaneous function used in main routines.
}
\examples{
y = c(1,2,3,4,1,2,3,4,1,2,3,4)
}

