% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{normalize}
\alias{normalize}
\title{A simple function to normalize a dataset of variables to the 0-1 interval.}
\usage{
normalize(X)
}
\arguments{
\item{X}{Matrix or dataframe to be normalized.}
}
\value{
A dataframe with normalized values ranging from 0 to 1.
}
\description{
This function returns a dataframe with normalized variables.
}
\examples{
someVector = 1:10
normalize(someVector)
df <- data(ITdata)
normalizedIT <- normalize(df)
describe(normalizedIT)
}

