% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{oftype}
\alias{oftype}
\title{Miscelaneous function used in main routines.}
\usage{
oftype(bestfit)
}
\arguments{
\item{bestfit}{a model object.}
}
\value{
either "lm", "logit", or "probit".
}
\description{
Miscelaneous function used in main routines.
}
\examples{
y=c(1,0,1,0,1,0)
x=c(0,0,1,1,1,0)
oftype(lm(y~x))
}

