% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{opt.glm}
\alias{opt.glm}
\title{Core function used by autoGLM, for details and usage see autoGLM.}
\usage{
opt.glm(data, reclasstable = "default", class = 1,
  outputpath = paste(getwd(), "//", sep = ""), modelname = "autoGLM",
  tracelevel = 1, actions = c("print", "return"), NAval = -9999,
  model = "logit", preselect = "lm", method = "opt.ic", crit.t = 1.64,
  crit.p = 0.1, test = "LR", KLIC = "AICc", accuracytolerance = 0.025,
  confidence.alternative = 0.85, use.share = 0.25, maxsampleruns = 50,
  memorymanagement = TRUE)
}
\arguments{
\item{data}{See also \code{\link{autoGLM}}.}

\item{reclasstable}{\code{\link{autoGLM}}.}

\item{class}{\code{\link{autoGLM}}.}

\item{outputpath}{\code{\link{autoGLM}}.}

\item{modelname}{\code{\link{autoGLM}}.}

\item{tracelevel}{\code{\link{autoGLM}}.}

\item{actions}{\code{\link{autoGLM}}.}

\item{NAval}{\code{\link{autoGLM}}.}

\item{model}{\code{\link{autoGLM}}.}

\item{preselect}{\code{\link{autoGLM}}.}

\item{method}{\code{\link{autoGLM}}.}

\item{crit.t}{\code{\link{autoGLM}}.}

\item{crit.p}{\code{\link{autoGLM}}.}

\item{test}{\code{\link{autoGLM}}.}

\item{KLIC}{\code{\link{autoGLM}}.}

\item{accuracytolerance}{\code{\link{autoGLM}}.}

\item{confidence.alternative}{\code{\link{autoGLM}}.}

\item{use.share}{\code{\link{autoGLM}}.}

\item{maxsampleruns}{\code{\link{autoGLM}}.}

\item{memorymanagement}{\code{\link{autoGLM}}.}
}
\value{
a model object.
}
\description{
Core function used by autoGLM, for details and usage see autoGLM.
}
\examples{
randomlogit <- simulateLogit(nobs=5000, pars = c(0.5, -0.4, -0.3, 0.1, 0, 0, 0, 0, 0, 0))
opt.glm(randomlogit, outputpath="C:\\\\users\\\\")
}

