% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{opt.h}
\alias{opt.h}
\title{Optimization routine based on hypothesis testing.}
\usage{
opt.h(model, Y, X, returntype, tracelevel, crit.p, test, memorymanagement)
}
\arguments{
\item{model}{The model to be optimized. Supports "lm" for the linear probability model, "logit" for the logistic probability model, and "probit" for the probit model.}

\item{Y}{The binary response variable.}

\item{X}{A dataframe with collumns of exogenous regressors.}

\item{returntype}{"model", "data", or "colnames"}

\item{tracelevel}{level of printing.}

\item{crit.p}{p value used in hypothesis tests.}

\item{test}{type of test, either "LR", "F", or "Chisq".}

\item{memorymanagement}{logical, indicating whether memory should be more actively managed.}
}
\value{
"model", "data", or "colnames", to be specified in returnype.
}
\description{
This function searches
}

