% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{opt.ic}
\alias{opt.ic}
\title{Optimization routine based on information criteria.}
\usage{
opt.ic(model, Y, X, KLIC = "AICc", returntype = "model", tracelevel = 0,
  memorymanagement = FALSE)
}
\arguments{
\item{model}{The model to be optimized. Supports "lm" for the linear probability model, "logit" for the logistic probability model, and "probit" for the probit model.}

\item{Y}{The binary response variable.}

\item{X}{A dataframe with collumns of exogenous regressors.}

\item{KLIC}{information criterion to be used, "AIC" or "AICc", See also \code{\link{IC}}}

\item{returntype}{"model", "data", or "colnames"}

\item{tracelevel}{level of printing.}

\item{memorymanagement}{logical, indicating whether memory should be more actively managed.}
}
\value{
"model", "data", or "colnames", to be specified in returnype.
}
\description{
This function optimizes a model using information critera as a decision rule. This solves many of the problems related to model selection based on hypothesis tests, see also \code{\link{opt.h}}.
In finite samples, the AIC is known to favor large models. The corrected AIC is a slightly stricter measure.
}
\examples{
randomlogit <- simulateLogit(nobs=500, pars = c(0.5, -0.4, -0.3, 0.1, 0.05, 0.025, 0, 0, 0, 0))
Y=randomlogit[,1]
X=randomlogit[,-1]
opt.ic(model="lm", Y, X)
}

