% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{pkgTest}
\alias{pkgTest}
\title{An function to load and install packages.}
\usage{
pkgTest(package, silent = FALSE)
}
\arguments{
\item{package}{Character string indicating th package to be loaded. Accepts a vector of packages to load multiple packages.}

\item{silent}{Should startup messages printed by the packages be surpressed? Will not surpress warnings or errors. Defaults to FALSE.}
}
\value{
The messages printed by the loaded package, or a message if the package is not found.
}
\description{
This function checks if a package is installed before loading it, and pulls it from the nearest CRAN server if the package is missing.
}
\examples{
verbose_packages <- c("gmm", "foreign", "sp")
silent_packages <- c("data.table", "compiler", "speedglm")

pkgTest(verbose_packages)
pkgTest(silent_packages)
}
\keyword{loading}
\keyword{package}

