% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{PtoBin}
\alias{PtoBin}
\title{Miscelaneous function used in main routines.}
\usage{
PtoBin(Pmap)
}
\arguments{
\item{Pmap}{a vector of probabilities.}
}
\value{
A binary vector indicating occurrences.
}
\description{
Miscelaneous function used in main routines.
}
\examples{
probabilities<-c(1:100)/100
occurrences<-PtoBin(probabilities)
}

