% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{selectX}
\alias{selectX}
\title{A main function of the package for variable selection based on moidel type and a generalize to specific approach.}
\usage{
selectX(Y, X, model = "lm", returntype = "data", method = "opt.ic",
  KLIC = "AICc", crit.t = 1.64, crit.p = 0.05, test = "LR",
  share = 0.75, confidence.alternative = 0.85, max.iter = 50,
  tracelevel = 1, memorymanagement = TRUE)
}
\arguments{
\item{Y}{A binary response variable.}

\item{X}{A dataframe of multiple exogenous regressors.}

\item{model}{Either "lm" for the linear probability model, "logit" for the logistic probability model, or "probit", for the probit model. The logit and probit models are solved using Iterated Weighted Least Squares, and optimization of the logit model is significantly faster than the probit model. Defaults to "lm".}

\item{returntype}{Either "data" to return a dataset, or colnames" to only return the collumn names of the variables that are used in the optimal model. "data" by default.}

\item{method}{The optimization strategy. Either "opt.ic" to optimize using information criteria, "opt.t" for step-wise elimination of insignificant values
(statistically speaking not a sound procedure, but it will provide a parsimonious model that can be usefull as a benchmark), or "opt.h" to optimize by classical hypothesis tests.
defaults to "opt.ic".}

\item{KLIC}{the information criterion used by "opt.ic", either "AIC" or "AICc", defaults to the latter.}

\item{crit.t}{The t-value indicating significance when using method "opt.t", defaults to 1.64.}

\item{crit.p}{the p-value used by method "opt.h" in the hypothesis tests. Defaults to 0.05.}

\item{test}{The hypothesis test used by "opt.h". Defaults to "LR" for the Likelihood Ratio test. Other options are "F", for an F test for joint significance of insignificant parameters, or "Chisq" for a wald test against the Chi squared distribution.
Recommended setting is either "LR" as it is less dependent on correct estimation of the standard errors. Keep in mind that "Chisq" is an asymptotic test, anf "F" is more appropiate for small sample tests. Howver "Chisq" holds under milder conditions and should be used if no small sample theory is available for the model.}

\item{share}{between 0-1, specifying the amount of data that should be passed on to the optimization strategies. Defaults to 0.75, to improve speed. Uses getSamples() to maintain first and second moments of the data.}

\item{confidence.alternative}{passed on to getSample. Defaults to .85.}

\item{max.iter}{passed on to getSample. Defaults to 50.}

\item{tracelevel}{the amount of information to be printed. Passed on to underlying routines. Defaults to 1 for printing, set to 0 for no printing.}

\item{memorymanagement}{TRUE/FALSE indicating whether garbage collection should be forec regularly when memory usage is high. Defaults to TRUE, recommended setting for large datasets.}
}
\value{
Either a dataframe of exogenous variables, or a vector containing the collumn names indicating the optimal variables extracted from the supplied dataset.
}
\description{
This function is a wrapper to the functions bestlinearX(), bestlogitX() and bestprobitX(), with an additional option to call getSamples for improved speed. Take into account that sampling itself takes time, such that total computational burden is a trade-off between the load of the getSample function and the model optimization itself.
}
\examples{
# load data
data(ITdata)
data(corinetable)
#Grab a sample (optional).
sample <- ITdata[getSamples(ITdata, share =.05),]
# Reclassify
catITdata <- reclassify(sample, reclasstable = corinetable)
# create a binary response dataset.
Y <- MLtoBinomData(catITdata[,1], class =1)
X <- catITdata[,-1]
selectX(Y, X, model ="lm", returntype = "colnames", method = "opt.t")
bestX <- selectX(Y, X)
describe(bestX)
}

