% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{simulateLogit}
\alias{simulateLogit}
\title{A function to generate random data from a logit model.}
\usage{
simulateLogit(nobs = 2500, pars = c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0))
}
\arguments{
\item{nobs}{number of observations that the returned dataset should include. Defaults to 2500.}

\item{pars}{A vector of parameters to be used in the data generating process. Regressors may have a 0 coefficient. Defaults to 0 for all variables.}
}
\value{
A simulated dataset from a logistic probability model with nobs observations, a binary response variable, and nxregs regressors ranging from 0 to 100.
}
\description{
This function generates random data simulated from a logit model based on the parameters supplied.
}
\examples{
logitpars <- c(-0.5, 0.4, -0.3, 0.2, -0.1, 0, 0, 0, 0, 0, 0, 0)
logit_data <- simulateLogit(nobs=2500, pars=logitpars)

random_binaryResponse_data <- simulateLogit(nobs=500)
}
\keyword{data}
\keyword{generate}
\keyword{logit}
\keyword{random}

