% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{update.start}
\alias{update.start}
\title{Miscelaneous function used in main routines.}
\usage{
\method{update}{start}(bestfit)
}
\arguments{
\item{bestfit}{modelobject.}
}
\value{
vector of starting values
}
\description{
Miscelaneous function used in main routines.
}

