% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{update.X}
\alias{update.X}
\title{Miscelaneous function used in main routines.}
\usage{
\method{update}{X}(bestfit, newX)
}
\arguments{
\item{bestfit}{model object}

\item{newX}{dataframe.}
}
\value{
dataframe
}
\description{
Miscelaneous function used in main routines.
}

