% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{update.Z}
\alias{update.Z}
\title{Miscelaneous function used in main routines.}
\usage{
\method{update}{Z}(Y, newX)
}
\arguments{
\item{Y}{vector}

\item{newX}{dataframe with corresponding rows.}
}
\value{
dataframe
}
\description{
Miscelaneous function used in main routines.
}

