% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{warmstart}
\alias{warmstart}
\title{Miscelaneous function used in main routines.}
\usage{
warmstart(oldstart, warmth = 1)
}
\arguments{
\item{oldstart}{vector of starting values.}

\item{warmth}{numeric between 0-1, to be multiplied by the supplied vector oldstart.}
}
\value{
vector of starting values
}
\description{
Miscelaneous function used in main routines.
}

