% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{withinInterval}
\alias{withinInterval}
\title{A function that determines whether a point lies in a specified interval.}
\usage{
withinInterval(x, interval)
}
\arguments{
\item{x}{numeric.}

\item{interval}{vector of length 2 specifying the outer values of the interval.}
}
\value{
TRUE/FALSE
}
\description{
A function that determines whether a point lies in a specified interval.
}
\examples{
withinInterval(5,c(1,10))
}

