% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\docType{data}
\name{ITdata}
\alias{ITdata}
\title{A sample dataset produced with \code{\link{getSamples}} with 157481 obersvations representing 100x100 meter resolution grid cells of Corine land cover data for Italy with 26 exogenous variables. Variable names starting with a "w" indicate spatial lags.}
\format{a data frame with 157481 rows and 27 columns}
\usage{
data(ITdata)
}
\description{
To inspect the data, describe(data(ITdata))
}
\keyword{datasets}

