% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{logistic}
\alias{logistic}
\title{The logistic function.}
\usage{
logistic(theta, data)
}
\arguments{
\item{theta}{A vector of coefficients.}

\item{data}{A dataframe of multiple exogenous regressors.}
}
\value{
A vector of values produced by a logistic formula under specified parameters and data.
}
\description{
Called by various routines in the package. May also be used to predict fitted conditional probabilities by supplying a set of variables and corresponding coefficients estimated from a logit model.
}
\examples{
logitdata <- simulateLogit(1000, c(1,0.5,-0.5,-0.3))
model <- logit(logitdata)
pars <- coef(model)
# predict with logistic function
predicted <- logistic(pars, cbind(1,logitdata[,-1]))
# compare with data
describe(logitdata[,1])
# compare with predict function from R
describe(fitted(model))
}

