% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{opt.glm}
\alias{opt.glm}
\title{Core function used by autoGLM, for details and usage see autoGLM.}
\usage{
opt.glm(data, reclasstable, class, outputpath, modelname, tracelevel = 1,
  actions = c("write", "print", "log", "return"), NAval = -9999,
  model = "logit", preselect = "lm", method = "opt.ic", crit.t = 1.64,
  crit.p = 0.1, test = "LR", KLIC = "AICc", accuracytolerance = 0.01,
  confidence.alternative = 0.9, use.share = 0.25, maxsampleruns = 100,
  memorymanagement = TRUE)
}
\arguments{
\item{data}{See also \code{\link{autoGLM}}.}

\item{reclasstable}{\code{\link{autoGLM}}.}

\item{class}{\code{\link{autoGLM}}.}

\item{outputpath}{\code{\link{autoGLM}}.}

\item{modelname}{\code{\link{autoGLM}}.}

\item{tracelevel}{\code{\link{autoGLM}}.}

\item{actions}{\code{\link{autoGLM}}.}

\item{NAval}{\code{\link{autoGLM}}.}

\item{model}{\code{\link{autoGLM}}.}

\item{preselect}{\code{\link{autoGLM}}.}

\item{method}{\code{\link{autoGLM}}.}

\item{crit.t}{\code{\link{autoGLM}}.}

\item{crit.p}{\code{\link{autoGLM}}.}

\item{test}{\code{\link{autoGLM}}.}

\item{KLIC}{\code{\link{autoGLM}}.}

\item{accuracytolerance}{\code{\link{autoGLM}}.}

\item{confidence.alternative}{\code{\link{autoGLM}}.}

\item{use.share}{\code{\link{autoGLM}}.}

\item{maxsampleruns}{\code{\link{autoGLM}}.}

\item{memorymanagement}{\code{\link{autoGLM}}.}
}
\value{
a model object.
}
\description{
Core function used by autoGLM, for details and usage see autoGLM.
}

