% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{opt.ic}
\alias{opt.ic}
\title{Optimization routine based on information criteria.}
\usage{
opt.ic(model, Y, X, KLIC, returntype, tracelevel, memorymanagement)
}
\arguments{
\item{model}{The model to be optimized. Supports "lm" for the linear probability model, "logit" for the logistic probability model, and "probit" for the probit model.}

\item{Y}{The binary response variable.}

\item{X}{A dataframe with collumns of exogenous regressors.}

\item{KLIC}{information criterion to be used, "AIC" or "AICc", See also \code{\link{IC}}}

\item{returntype}{"model", "data", or "colnames"}

\item{tracelevel}{level of printing.}

\item{memorymanagement}{logical, indicating whether memory should be more actively managed.}
}
\value{
"model", "data", or "colnames", to be specified in returnype.
}
\description{
This function searches
}
\examples{
opt.ic(model, Y, X, KLIC, returntype, tracelevel, memorymanagement)
}

