% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGLM.R
\name{opt.t}
\alias{opt.t}
\title{Optimization routine based on step-wise elimenation using t-values.}
\usage{
opt.t(model, Y, X, returntype = "model", tracelevel = 0, crit.t = 1.64,
  memorymanagement = FALSE)
}
\arguments{
\item{model}{The model to be optimized. Supports "lm" for the linear probability model, "logit" for the logistic probability model, and "probit" for the probit model.}

\item{Y}{The binary response variable.}

\item{X}{A dataframe with collumns of exogenous regressors.}

\item{returntype}{"model", "data", or "colnames"}

\item{tracelevel}{level of printing.}

\item{crit.t}{t-value used for significance test.}

\item{memorymanagement}{logical, indicating whether memory should be more actively managed.}
}
\value{
"model", "data", or "colnames", to be specified in returnype.
}
\description{
This function searches
}
\examples{
randomlogit <- simulateLogit(nobs=500, pars = c(0.5, -0.4, -0.3, 0.1, 0.05, 0.025, 0, 0, 0, 0))
Y=randomlogit[,1]
X=randomlogit[,-1]
opt.t(model="lm", Y, X)
}

